sidebar permission && crud in user

+++ resources/views/users/index.blade.php
@@ -10,9 +10,9 @@
           <h1>Users Management</h1>
         </div>
         <div class="col-sm-6">
-
+            @can('user-create')
             <a class="btn btn-success float-right" href="{{ route('users.create') }}"> Create New User</a>
-
+            @endcan
         </div>
 
       </div>
@@ -33,7 +33,7 @@
             <th style="width: 10px">#</th>
             <th>Name</th>
             <th>Email</th>
-            <th>Rolles</th>
+            <th>Roles</th>
             <th>Action</th>
           </tr>
         </thead>
@@ -63,17 +63,19 @@
               </td>
 
               <td>
+                  @can('user-list')
+                      <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+                  @endcan
 
-                 <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+                  @can('user-edit')
+                      <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+                  @endcan
 
-                 <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-
-                  {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-
-                      {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-
-                  {!! Form::close() !!}
-
+                  @can('user-delete')
+                       {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                       {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                     {!! Form::close() !!}
+                  @endcan
               </td>
 
             </tr>
